04c-conf_ensemble_masif_precompute.py - new script updated - for handling and precomputation of np files (e.g. x,y,z coordinates, polar coordinates, iface labels, input featrues, etc) of configurtaions from structureal ensemble. The ensemble is generated by HTMD & clustered by CLoNe

+++ source/data_preparation/01d-conf_ensemble_pdb_extract_and_triangulate.py
@@ -180,4 +180,5 @@ for i in np.arange(1,cluster_num+1):
         shutil.copy(out_filename1+'.ply', ply_chain_dir) 
         shutil.copy(out_filename1+'.pdb', pdb_chain_dir) 
         # Clean up the tmp pdb file
-        os.remove(pdb_filename)
+        if os.path.isfile(pdb_filename):
+            os.remove(pdb_filename)
